requirements.txt.txt
streamlit
pandas
pybase64
pdfminer.six
python-docx
Pillow
matplotlib